requirements.txt
fastapi==0.124.4
uvicorn[standard]==0.38.0
httpx>=0.27.0
aiosqlite>=0.19.0
pywinrm==0.5.0
pydantic>=2.5.0,<3.0
python-jose[cryptography]==3.3.0
python-dotenv>=1.0.0
cloudflare>=4.3.0
bcrypt>=4.0.0
slowapi>=0.1.9
requests-ntlm>=1.2.0
requests>=2.31.0
impacket>=0.11.0

